pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dev-reliability-env"
version = "1.0.0"
description = "CI debugging and SRE incident response environment for OpenEnv."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",
    "openenv-core",
    "pydantic>=2.7.0",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.0",
]

[tool.setuptools]
py-modules = ["client", "models"]

[tool.setuptools.packages.find]
include = ["server", "tasks", "tasks.*", "tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
